Refactor `DraftModeBanner` by removing `primaryNavigation` support and restructuring its file path.
