+++ app/layout.tsx
@@ -10,7 +10,7 @@ import "@uoguelph/web-components/style";
 import { Button } from "@uoguelph/react-components/button";
 import { WebComponentsLoader, WebComponentsLoaderProps } from "@/components/client/web-components-loader";
 import { BackToTop } from "@uoguelph/react-components/back-to-top";
-import { DraftModeBanner } from "@/components/client/draft-mode-banner";
+import { DraftModeBanner } from "@/components/client/draft-mode/draft-mode-banner";
 
 const dmSans = DM_Sans({
   subsets: ["latin"],
